chore(ci): setup CI
This commit adds a set of GH workflows and updates pre-commit with new
hook (`yamllint`).

Note that the workflows were shamelessly stolen from jonhoo's Rust CI
repo configs. One important difference is that as here we have no tests
yet, the code coverage job was not ported.

Also, Windows build is currently bugged as the `xdg` crate is only for
unix systems. Therefore, the Windows CI test is temporarily disabled
until this is fixed.