feat(cache): add basic bitlink cache
This commit implements basic version of a bitlink cache that is backed
by a local SQLite database.

The database (driver) is bundled with the CLI app, so no additional
external dependency is required.

The cache directory can be adjusted by either:
 - Adding a `cache_dir` to the TOML config file
 - Specifying an ad-hoc `--cache-dir` CLI option

If left unspecified, the default cache directly follows the XDG
directory convention (so most likely `~/.cache/bitcli/`).

Note that an empty directory path disables the cache and that the
concrete `.db` filename is based on the API version.

Finally, note that the cache DB is _not_ encrypted, so it's currently
advised to disable it if any of the potential URLs could be considered
as a sensitive information.