feat(cli): support offline mode with `--offline`
This commit adds new CLI flag `--offline`. Under the _offline_ mode, the
API client won't issue any HTTP requests and will only rely on the local
cache.

In case the cache does not contain requested data, an error is reported.

Note that the CLI option cannot be combined with `--no-cache` (for
obvious reasons) and that disabling the cache via `--cache-dir` will in
turn disable the offline mode.