feat(cli): add `--ordering` option for shorten
This commit adds new option `--ordering=<ordered|unordered>` for the
`shorten` command.

The ordering variants specify whether the output follows the order of
URLs given as arguments (`ordered`) or not (`unordered`). Preserving the
input order is the default, however, the allowing arbitrary order may be
more efficient (especially with larder values of `--max-concurrent`).

Another difference is in the output format. Since it would be difficult
to match the shortened links of `unordered` back to the inputs, with
this option the output additionally includes the long URLs at each line.

 - Ordered output line: `<short-url>\n`
 - Unordered output line: `<short-url>\t<long-url>\n`