feat(cli): read URLs to shorten from stdin
This commit adds support for reading (lazily) URLs to be shortened from
stdin when none is given as in the argument list.

The standard input is expected to yield one URL per line and currently
is expected to be non-empty. That is, the application will wait for at
least one input line indefinitely. This might be leveraged later.

Also note that the program will terminate on the first URL parsing
error and in that case the remaining input won't be processed (i.e., no
further API requests will be issued).